Move text tools into a dedicated submodule
