+++ elgeopaso/jobs/analyzer/__init__.py
@@ -1,4 +1,3 @@
 #! python3  # noqa: E265
-# -*- coding: utf-8 -*-
 
 from .georezo.georezo_analyzer import GeorezoOfferAnalizer  # noqa: F401
