+++ elgeopaso/jobs/analyzer/georezo/georezo_analyzer.py
@@ -35,6 +35,7 @@ from elgeopaso.jobs.models import (
     Technology,
     TechnologyVariations,
 )
+from elgeopaso.utils import TextToolbelt
 
 from .parsers import ContentParser, TitleParser
 
@@ -45,6 +46,8 @@ from .parsers import ContentParser, TitleParser
 # timestamps format helpers
 _regex_markups = re.compile(r"<.*?>|&([a-z0-9]+|#[0-9]{1,6}|#x[0-9a-f]{1,6});")
 
+# shortcuts
+txt_toolbelt = TextToolbelt()
 
 # ############################################################################
 # ########## Classes ##############
@@ -107,22 +110,21 @@ class GeorezoOfferAnalizer:
             raw_offer = GeorezoRSS.objects.get(id_rss=offer_id)
 
             # -- Title analisis ----------------------
-            clean_title = self.remove_html_markups(raw_offer.title)
+            clean_title = txt_toolbelt.remove_html_markups(raw_offer.title)
             title_parser = TitleParser(offer_id=offer_id, input_title=clean_title)
 
             # determine contract type
             contract_type = title_parser.parse_contract_type()
+            jobs_labels = title_parser.parse_jobs_positions()
             place = title_parser.parse_place(mode=0)
 
             # -- Content analisis ----------------------
-            clean_content = self.remove_html_markups(raw_offer.content)
+            clean_content = txt_toolbelt.remove_html_markups(raw_offer.content)
 
             content_parser = ContentParser(
                 offer_id=offer_id, input_content=clean_content
             )
-            content_title = self.parse_words(raw_offer.title)
             technos = content_parser.parse_technology()
-            jobs_labels = self.parse_jobs_positions(content_title)
 
             # add or update offer
             if self.new:
@@ -171,153 +173,6 @@ class GeorezoOfferAnalizer:
             clean_offer.technologies.set(technos)
             clean_offer.jobs_positions.set(jobs_labels)
             logging.debug("Offer analyzed and inserted jobs.offer: {}".format(offer_id))
-
-    # PARSERS ----------------------------------------------------------------
-    def parse_words(self, offer_raw_content):
-        """
-        Extraction of words mentioned into the offers. The goal is to perform
-        a semantic analysis.
-        It's based on NLTK: https://www.nltk.org/
-        """
-        # get list of common French words to filter
-        stop_fr = set(stopwords.words("french"))  # add specific French
-
-        # custom list
-        li_stop_custom = (
-            "(",
-            ")",
-            "...",
-            ".",
-            ":",
-            ";",
-            "/",
-            "nbsp",
-            "&",
-            "#",
-            ",",
-            "-",
-            ":",
-            "http",
-            "img",
-            "br",
-            "amp",
-            "<",
-            ">",
-            "%",
-            "border",
-            "*",
-            "border=",
-            "les",
-            "leurs",
-            "&",
-            "#",
-            "-",
-            "+",
-            ":",
-            ".",
-            ";",
-            "à",
-            "où",
-            "des",
-            ",",
-            "nbsp",
-            "De",
-            "Des",
-            "et",
-            "en",
-            "(",
-            ")",
-            "pour",
-            "plus",
-            "sein",
-            "sous",
-            "Les",
-            "auprès",
-            "etc",
-            "the",
-            "for",
-            "ème",
-            "via",
-            "Vos",
-            "dès",
-            "plein",
-            "tel",
-            "etc.",
-            "etc..",
-            "Ces",
-            "tél",
-            "cela",
-            "ceci",
-            "cet",
-        )
-
-        contenu = BeautifulSoup(offer_raw_content, "html.parser")
-        contenu = contenu.get_text("\n")
-        contenu = self.remove_html_markups(offer_raw_content)
-        # contenu = self.clean_xml(contenu)
-        contenu_tokenized = nltk.word_tokenize(contenu)
-        # print(len(contenu_tokenized))
-
-        # stop words filter
-        for mot in contenu_tokenized:
-            if mot in stop_fr or mot in li_stop_custom:
-                contenu_tokenized = list(filter((mot).__ne__, contenu_tokenized))
-
-        logging.debug("Words parsed: {}".format(len(contenu_tokenized)))
-
-        # print(len(contenu_tokenized))
-        return contenu_tokenized
-
-    def parse_technology(self, offer_content_tokenized):
-        """TO DOCUMENT
-        """
-        technos_matched = []
-        # print(sorted(offer_content_tokenized))
-        for word in offer_content_tokenized:
-            if TechnologyVariations.objects.filter(label=word.lower()).exists():
-                techno_name = TechnologyVariations.objects.get(label=word.lower()).name
-                technos_matched.append(Technology.objects.get(name=techno_name))
-            else:
-                continue
-        logging.debug("Technologies identified: {}".format(technos_matched))
-        return technos_matched
-
-    def parse_jobs_positions(self, offer_content_tokenized):
-        """TO DOCUMENT
-        """
-        jobs_positions_matched = []
-        for word in offer_content_tokenized:
-            if JobPositionVariations.objects.filter(label=word.lower()).exists():
-                job_label = JobPositionVariations.objects.get(label=word.lower()).name
-                jobs_positions_matched.append(JobPosition.objects.get(name=job_label))
-            else:
-                continue
-        logging.debug("Jobs positions identified: {}".format(jobs_positions_matched))
-        return jobs_positions_matched
-
-    # ------------ UTILITIES -------------------------------------------------
-    @classmethod
-    def remove_html_markups(cls, html_text: str, cleaner: str = "bs-lxml") -> str:
-        """Very basic cleaner for HTML markups.
-
-        :param str html_text: text to be clean
-        :param str cleaner: Which lib to use to clean the text:
-          - "bs-lxml": Beautifulsoup4 + LXML - Default.
-          - "psl-only": Python Standard Library only (html + regex)
-
-        :return: clean text
-        :rtype: str
-        """
-        # with BeautifulSoup + LXML
-        if cleaner == "bs-lxml":
-            cleaned_text = BeautifulSoup(html_text, "lxml").text
-        elif cleaner == "psl-only":
-            # convert HTML5 characters into str.
-            # See: https://docs.python.org/3/library/html.html#html.unescape
-            html_text = html.unescape(html_text)
-            cleaned_text = _regex_markups.sub(" ", html_text)
-
-        return cleaned_text
 
 
 # ############################################################################
